research-cohort/requirements.txt
gradio
openai
python-dotenv
aiohttp
